+++ lib/screens/doctor/doctor_home.dart
@@ -3,15 +3,15 @@ import 'package:patientpulse/backend/patients.dart';
 import 'package:patientpulse/extensions/extensions.dart';
 import 'package:patientpulse/main.dart';
 
-class PatientHome extends StatelessWidget {
-  const PatientHome({super.key});
+class DoctorHome extends StatelessWidget {
+  const DoctorHome({super.key});
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
         backgroundColor: Color.fromARGB(255, 5, 47, 82),
-        title: Text('PatientPulse').color(Colors.white),
+        title: Text('PatientPulse (admin)').color(Colors.white),
         centerTitle: true,
         actions: [
           Icons.logout.toIcon(color: Colors.white).onClick(() async {